Breaks uitls tests into modules
